Refactor vertical whitespace and comment tracking into AST.
Previously vertical whitespace would need to be worked out 'by hand' using the
offsets of AST nodes and inspecting the input string. Comments also would just
be collected into a separate list and, again, manually reconciled with the AST.

While comments and whitespace are not semantically important, they are
important when performing translations since preserving this semantic detail is
essential. This change makes it easier to access this information.